feat: withdraw

+++ migrations/000002_create_orders_table.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
     number VARCHAR(255) UNIQUE NOT NULL,
     status order_status NOT NULL,
-    accrual INTEGER,
+    accrual NUMERIC,
     uploaded_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     user_id INTEGER NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
